+++ tests/__init__.py
@@ -1,1 +1,5 @@
-from .rare_tests import RareAPITests
+from .rareuser_tests import RareUserAPITests
+from .post_tests import PostTests
+from .category_tests import CategoryTests
+from .tag_tests import TagTests
+from .comment_tests import CommentTests
